--- src/main/java/ladder/LadderGame.java
@@ -1,7 +1,0 @@
-package ladder;
-
-public class LadderGame {
-    public static LadderGameResults start(Ladder ladder, LadderGameInfo ladderGameInfo) {
-        return ladderGameInfo.match(ladder.exec());
-    }
-}
