+++ src/main/java/ladder/domain/DirectionCreator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public interface DirectionCreator {
     boolean create();
