+++ src/main/java/ladder/domain/LadderCreator.java
@@ -1,4 +1,6 @@
-package ladder;
+package ladder.domain;
+
+import ladder.dto.LadderInfo;
 
 import java.util.List;
 import java.util.stream.Collectors;
