+++ src/main/java/ladder/domain/LadderResults.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.Map;
 import java.util.Objects;
