+++ src/main/java/ladder/domain/LadderReward.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.Objects;
