+++ src/test/java/ladder/DirectionTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.domain.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
