+++ src/test/java/ladder/LadderCreatorTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.domain.*;
+import ladder.dto.LadderInfo;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
