+++ src/test/java/ladder/LadderResultsTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.domain.LadderResults;
+import ladder.domain.Position;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
