+++ src/test/java/ladder/LadderTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
