+++ src/test/java/ladder/UserTest.java
@@ -1,7 +1,7 @@
 package ladder;
 
+import ladder.domain.User;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
