+++ src/test/java/ladder/UsersTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.domain.User;
+import ladder.domain.Users;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
